# Workast_JWT/api/swagger.yml
swagger: "2.0"
info:
  description: "Workplace Test Api"
  version: "1.0.0"
  title: "workplace"
  contact:
    email: "pilarcastro86@gmail.com"
host: "localhost:5001"
basePath: "/api/v1"
tags:
- name: "Users"
  description: "Informacion sobre los usuarios"
- name: "Articles"
  description: "Articulos publicados"
schemes:
- 'http'
paths:
  '/users':
    post:
      tags:
      - "Users"
      summary: "Crear un nuevo usuario en la db"
      operationId: "postUsers"
      consumes:
        - application/x-www-form-urlencoded
      parameters:
      - in: "body"
        name: "user"
        description: "Usuario a agregar a la db"
        required: true
        schema:
          type: object
          properties:
            name:
              type: string
            avatar:
              type: string
      responses:
        405:
          description: "Invalido"
        200:
          description: "Se creo un nuevo usuario"

  
  

    get:
      tags:
      - "Users"
      summary: "Mostrar todos los usuarios de la bd"
      operationId: "getUsers"
      consumes:
      - "application/x-www-form-urlencoded"
      produces:
      - "application/xml"
      - "application/json"
      responses:
        200:
          description: "successful operation"
          schema:
            type: object
            properties:
              name:
                type: string
              avatar:
                type: string
        400:
          description: "Invalid status value"
  
  
          
        
        
        
        
    put:
      tags:
      - "Users"
      summary: "Actualizar datos de un usuario"
      operationId: "updateUsers"
      consumes:
      - "application/json"
      - "application/xml"
      produces:
      - "application/xml"
      - "application/json"
      parameters:
      - in: "body"
        name: "user"
        description: "Usuario para modificar"
        required: true
        schema:
          type: object
          properties:
            name:
              type: string
            avatar:
              type: string
      responses:
        400:
          description: "ID de usuario invalido"
        404:
          description: "No se encontro el usuario"
        200:
          description: "Se modifico el usaurio"
          
          
          
          
    delete:
      tags:
      - "Users"
      summary: "Eliminar Usuario"
      operationId: "deleteUser"
      produces:
      - "application/xml"
      - "application/json"
      parameters:
      - name: "UserId"
        in: "path"
        description: "Usuario a eliminar"
        required: true
        type: "integer"
      responses:
        400:
          description: "ID de usuario invalido"
        404:
          description: "Usuario no encontrado"
          
          
          
  '/articles':
    post:
      tags:
      - "Articles"
      summary: "Crear un nuevo articulo en la db"
      operationId: "postArticles"
      consumes:
        - application/x-www-form-urlencoded
      parameters:
      - in: "body"
        name: "user"
        description: "Articulo a agregar a la db"
        required: true
        schema:
          type: object
          properties:
            userId:
              type: string
            title:
              type: string
            text:
              type: string
            tags:
              type: array
              items:
                type: string
      responses:
        405:
          description: "Invalido"
        200:
          description: "Se creo un nuevo articulo"

  
  

    get:
      tags:
      - "Articles"
      summary: "Mostrar todos los Articulos de la bd"
      operationId: "getArticles"
      consumes:
      - "application/x-www-form-urlencoded"
      produces:
      - "application/xml"
      - "application/json"
      responses:
        200:
          description: "successful operation"
          schema:
            type: object
            properties:
              userId:
                type: string
              title:
                type: string
              text:
                type: string
              tags:
                type: array
                items:
                  type: string
        400:
          description: "Invalid status value"
  
  
          
        
        
        
        
    put:
      tags:
      - "Articles"
      summary: "Actualizar Articles"
      operationId: "updateArticles"
      consumes:
      - "application/json"
      - "application/xml"
      produces:
      - "application/xml"
      - "application/json"
      parameters:
      - in: "body"
        name: "user"
        description: "Articulo para editar"
        required: true
        schema:
          type: object
          properties:
            userId:
              type: string
            title:
              type: string
            text:
              type: string
            tags:
              type: array
              items:
                type: string
      responses:
        400:
          description: "ID de articulo invalido"
        404:
          description: "No se encontro el articulo"
        200:
          description: "Se modifico el articulo exitosamente"
          
          
          
          
    delete:
      tags:
      - "Articles"
      summary: "Eliminar Articulo"
      operationId: "deleteArticles"
      produces:
      - "application/xml"
      - "application/json"
      parameters:
      - name: "ArticleId"
        in: "path"
        description: "Articulo a eliminar"
        required: true
        type: "integer"
      responses:
        400:
          description: "ID de articulo invalido"
        404:
          description: "Articulo no encontrado"
  
  
  
securityDefinitions:
  api_key:
    type: "apiKey"
    name: "api_key"
    in: "header"